Update riscv ignore file.
The test_thread_dimensions, test_select, and test_bruteforce tests are
the only tests that take too long to run.

For test_thread_dimensions, we can limit it to 1D and 2D tests.
For test_select, we can run it in Wimpy mode.
For test_bruteforce, we can increase the Wimpy reduction factor.